+++ pkg/sfu/pacer/pacer.go
@@ -34,7 +34,7 @@ type Packet struct {
 	TransportWideExtID uint8
 	WriteStream        webrtc.TrackLocalWriter
 	Metadata           interface{}
-	OnSent             func(md interface{}, sentHeader *rtp.Header, payloadSize int, sentTime time.Time, sendError error)
+	OnSent             func(md interface{}, marker bool, hdrSize int, payloadSize int, sentTime time.Time, sendError error)
 }
 
 type Pacer interface {
